typedef uint32_t and int32_t

+++ src/structs.h
@@ -6,8 +6,8 @@ typedef unsigned char   u8;
 typedef signed   char   s8;
 typedef unsigned short  u16;
 typedef signed   short  s16;
-typedef unsigned long   u32;
-typedef signed   long   s32;
+typedef uint32_t  u32; 
+typedef int32_t   s32;
